Make providers_controller_spec work
1. Made an interface class that contains a hash of {prompt: input}
2. Made two different 'adapters', one takes a user input as the input, the
other takes a text input.
