+++ spec/controllers/providers_controller_spec.rb
@@ -1,6 +1,6 @@
-require_relative '../../controllers/providers_controller'
-
 require_relative '../../models/interface'
+require_relative '../../controllers/providers_controller'
+require 'pry'
 
 # RSpec.describe ProvidersController do
 #   it 'adds a provider to the list' do
@@ -14,16 +14,16 @@ RSpec.describe ProvidersController do
   it 'adds a provider to the list' do
     $all_providers = []
 
-    inputs = {
-      'Please enter a command:' => 'sp:add',
-      'Provider Name:' => 'jason',
-      'Provider Phone Number:' => '555'
-    }
+    text_input_strategy = TextInputStrategy.new([
+      { 'Please enter a command:' => 'sp:add' },
+      { 'Provider Name:' => 'Jason' },
+      { 'Provider Phone Number:' => '555' },
+      { 'Please enter a command:' => 'exit' },
+   ])
 
-    inputs.each do |prompt, input|
-      Interface.enter(input)
-    end
+    interface = Interface.new(text_input_strategy)
+    interface.start
 
-      expect(ProviderHelper.new.find_provider_by_name('Jason').name).to eq ('Jason')
-    end
+    expect(ProviderHelper.new.find_provider_by_name('Jason').name).to eq ('Jason')
+  end
 end
